Fix exception when parsing LID for non-stations
'L' key in the LocationID is present for stations.
But sometimes a non-station is return as a result of a location search.
Such non-stations have a unique ID in their 'b' key of their LID.

d.get(k) returns None if k is not in d, but d[k] throws exception.
I.e. if neither 'L' or 'b' is present, we still throw an exception.